% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/be_save.r
\name{be_save}
\alias{be_save}
\title{Custom settings for ggsave}
\usage{
be_save(plot = NULL, filename = "", dpi = 700, size_ppt = "quarter",
  height = 0, width = 0)
}
\arguments{
\item{No}{parameters}
}
\description{
Custom preset settings for ggsave, fitting for PPT presentations.
}
\examples{
be_save()
}
\keyword{ggplot,}
\keyword{ggthemes,}
\keyword{theme}

