% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cat2Prob.r
\name{Cat2Prob}
\alias{Cat2Prob}
\title{Function to calculate target-likelyhood for groups for a categorical value}
\usage{
Cat2Prob(varname, target, df, cutoff, roundprob)
}
\arguments{
\item{varname}{- variable name, target - target variable name, df - the data frame,
cutoff - the minimum N for a categorical value, roundprob - the number of groups to report
(rounded likelyhoods)}
}
\description{
This function is used to return the target-likelyhood for groups for a categorical value.
It is used if the variable contains to many groups to be used on its own.
}
\examples{
Cat2Prob()
}

