% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDenom.r
\name{FindDenom}
\alias{FindDenom}
\title{Function for finding the denominator of a normalized vector}
\usage{
FindDenom(input)
}
\arguments{
\item{input}{(EXPLAIN!)}
}
\description{
#This function takes a vector that has been normalized and tries to find the denominator
used for normalizing. To un-normalize the vector multiply it by the result of this function,
like this UnNorm <- Normalized*FindDenom(Normalized).
}
\details{
The function works by trying to find out how big a unit of 1 in the unnormalized set is in
this normalized set. We find this by sorting and checking the most common difference
(Norm[i]-Norm[i-1]). The most common difference probably represents the unit 1.
}
\examples{
FindDenom()
}
\keyword{denominator}
\keyword{normalization,}
\keyword{un-normalize,}

