% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftCharts.r
\name{liftCharts}
\alias{liftCharts}
\title{Plots lift and gain}
\usage{
liftCharts(actuals, predictions, groups = 10)
}
\arguments{
\item{actuals}{= the truth
predictions = the predictions
groups = number of groups (ex. deciles (10))}
}
\description{
Function for generating lift and gain charts, using actuals and predictions.
}
\keyword{charts,}
\keyword{dplyr,}
\keyword{gain,}
\keyword{ggplot}
\keyword{lift,}
\keyword{plots,}
