% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingPlot.r
\name{missingPlot}
\alias{missingPlot}
\title{Missing variable plot}
\usage{
missingPlot(data, prop = FALSE, numbers = TRUE, sortVars = TRUE,
  ylab = c("Missing data", "Pattern"), colors = c("lightgrey", "red"),
  gap = 3, cex.axis = 0.7)
}
\description{
Visualization of missing variables using VIM.
}
\keyword{VIM,}
\keyword{missing}
\keyword{variables,}
\keyword{visualization}
