% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monyr.r
\name{monyr}
\alias{monyr}
\title{Funtion for converting date to mon.yr}
\usage{
monyr(date_column)
}
\description{
Converts any date to format mon.yr (ex. Jul.17, Aug.14, etc), as an ordered factor.
}
\examples{
monyr(df$date)
}
\keyword{conversion}
\keyword{date}
