% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSave.r
\name{qSave}
\alias{qSave}
\title{Custom settings for ggsave}
\usage{
qSave(plot = NULL, filename = "", folder = "figs", dpi = 700,
  size_ppt = "quarter", height = 0, width = 0)
}
\arguments{
\item{plot}{= plot object (without quotes)
filename = filename including folder structure from working directory
dpi = dpi setting
size_ppt = size parameter, in parts of a powerpoint slide ("quarter", "half high", "half wide", or "full")
height = custom height
width = custom width}
}
\description{
Custom preset settings for ggsave, with custom sizes for PPT presentations.
}
\examples{
## After creating plot, use qSave() to save the plot with preset options

## Filename created from the supplied plot object name, folder "./Gfx"
## created in working directory, plot saved with default (quarter ppt) size.
qSave(plot = plt.test)

## Saved with custom file name, with size half heigh full width
qSave(plot = plt.test, filename = "Gfx/CUSTOM_NAME.png", size_ppt = "halfwide")

## Saved in custom folder, created if it doesn't already exist
qSave(plot = plt.test, folder = "CUSTOM_FOLDER")
}
\keyword{ggplot,}
\keyword{ggthemes,}
\keyword{theme}
