% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_be.r
\name{theme_be}
\alias{theme_be}
\title{A custom theme for ggplot2}
\usage{
theme_be(base_size = 15, axis.lines = FALSE, light.grid = TRUE,
  legend.position = "right", legend.background = element_blank(),
  legend.title = element_text())
}
\arguments{
\item{No}{parameters}
}
\description{
A simple custom theme setup, based on theme_bw, base size 15. The theme is quite minimal.
Use theme(axis.line.x = element_blank(), axis.line.y=element_blank()) to remove axis lines.
}
\examples{
ggplot(df, aes(x = x)) +
geom_histogram() +
theme_be()
theme_be()
}
\keyword{ggplot,}
\keyword{ggthemes,}
\keyword{theme}
